requirements.txt
pycryptodome==3.20.0
pygame==2.6.0
